Add a Nix package for roastery
Also add `rye` to the shell, so we can also build stuff with the normal
Python ecosystem for local development. The normal dev tools play a bit
nicer with editors such as PyCharm and also do not require rebuilding
the entire Nix derivation every time we make a small change.